Reshuffling code.
