+++ extensions/amp-ad-network-base/0.1/amp-ad-network-base.js
@@ -35,7 +35,7 @@ export class AmpAdNetworkBase extends AMP.BaseElement {
     super(element);
 
     /** @private {?Promise<!../../../src/service/xhr-impl.FetchResponse>} */
-    this.adPromise_ = null;
+    this.adResponsePromise_ = null;
 
     /** @private {Object<string, !./amp-ad-type-defs.Validator>} */
     this.validators_ = map();
@@ -70,6 +70,7 @@ export class AmpAdNetworkBase extends AMP.BaseElement {
       // TODO(levitzky) handle non-numeric values.
       width: this.element.getAttribute('width'),
       height: this.element.getAttribute('height'),
+      layout: this.element.getAttribute('layout'),
     };
   }
 
@@ -79,19 +80,19 @@ export class AmpAdNetworkBase extends AMP.BaseElement {
   }
 
   /** @override */
-  layoutCallback() {
-    dev().assert(this.adPromise_, 'layoutCallback invoked before XHR request!');
-    return this.adPromise_
-        .then(response => this.invokeValidator_(response))
-        .then(validatorResult => this.invokeRenderer_(validatorResult))
-        .catch(error => this.handleFailure_(error.type, error.msg));
-  }
-
-  /** @override */
   onLayoutMeasure() {
     this.sendRequest_();
   }
 
+  /** @override */
+  layoutCallback() {
+    dev().assert(this.adResponsePromise_,
+        'layoutCallback invoked before XHR request!');
+    return this.adResponsePromise_
+        .then(response => this.invokeValidator_(response))
+        .then(validatorResult => this.invokeRenderer_(validatorResult))
+        .catch(error => this.handleFailure_(error.type, error.msg));
+  }
 
   /**
    * @param {!./amp-ad-type-defs.FailureType} failure
@@ -131,11 +132,10 @@ export class AmpAdNetworkBase extends AMP.BaseElement {
   }
 
   /**
-   * @param {string} name
-   * @param {*} value
+   * @return {!Object} The context object passed to validators and renderers.
    */
-  setContextField(name, value) {
-    this.context_[name] = value;
+  getContext() {
+    return this.context_;
   }
 
   /**
@@ -159,7 +159,7 @@ export class AmpAdNetworkBase extends AMP.BaseElement {
   sendRequest_() {
     Services.viewerForDoc(this.getAmpDoc()).whenFirstVisible().then(() => {
       const url = this.getRequestUrl();
-      this.adPromise_ = sendXhrRequest(url, this.win);
+      this.adResponsePromise_ = sendXhrRequest(this.win, url);
     });
   }
 
@@ -169,7 +169,7 @@ export class AmpAdNetworkBase extends AMP.BaseElement {
    * @return {!Promise}
    * @private
    */
-  invokeValidator_(response) {
+  invokeValidator_(response) { debugger;
     if (!response.arrayBuffer) {
       return Promise.reject(this.handleFailure_(FailureType.INVALID_RESPONSE));
     }
